// src/types/convert.ts
export type ConvertType = 
  | 'pdftodoc' 
  | 'pdftoexcel' 
  | 'pdftoppt' 
  | 'pdftojpg'
  | 'doctopdf'
  | 'exceltopdf'
  | 'ppttopdf'
  | 'jpgtopdf'
  | 'ocrpdf'
  | 'compresspdf'
  | 'encryptpdf'
  | 'unlockpdf'
  | 'pdftoepub'
  | 'pdftohtml'
  | 'pdftotiff'
  | 'pdftotex'
  | 'pdftosvg'
  | 'pdftomobi'
  | 'pdftoxps'
  | 'pdftoxml'
  | 'epubtopdf'
  | 'htmltopdf'
  | 'textopdf'
  | 'pstopdf'
  | 'xslfotopdf'
  | 'xpstopdf'
  | 'svgtopdf'
  | 'pcltopdf'
  | 'xmltopdf'
  | 'mdtopdf';